app/src/main/java/com/vijaydhoni/newsapp/data/api/NewsAPiService.kt
package com.vijaydhoni.newsapp.data.api


import com.vijaydhoni.newsapp.BuildConfig
import com.vijaydhoni.newsapp.data.model.APIResponse
import retrofit2.Response
import retrofit2.http.GET
import retrofit2.http.Query

interface NewsAPiService {

    @GET("v2/top-headlines")
    suspend fun getTopHeadlines(
        @Query("country")
        country: String,
        @Query("page")
        page: Int,
        @Query("apiKey")
        apiKey: String = BuildConfig.API_KEY
    ): Response<APIResponse>


    @GET("v2/top-headlines")
    suspend fun getSearchedHeadlines(
        @Query("country")
        country: String,
        @Query("q")
        searchQuery: String,
        @Query("page")
        page: Int,
        @Query("apiKey")
        apiKey: String = BuildConfig.API_KEY
    ): Response<APIResponse>


}